agent_server/requirements.txt
aiohttp>=3.9.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
sentence-transformers>=2.2.2
pymupdf>=1.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0